# openapi/gateway.yaml
...
openapi: 3.0.3
info:
  title: SkyLink API
  version: 0.6.2
  description: |
    SkyLink Connected Aircraft Platform - MVP

    Security by Design implementation with JWT authentication,
    rate limiting, and strict input validation.
  contact:
    name: SkyLink Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.skylink.example.com
    description: Production server
...
paths:
...
  /contacts/oauth/callback:
    get:
      tags: [contacts]
      summary: OAuth callback (exchange code for tokens)
      description: |
        Exchanges `code` for Google tokens (access + refresh) and associates them with the application account.
        Redirects to frontend if `redirect_uri` was provided at initiation.
      operationId: googleOAuthCallback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code returned by Google.
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State value provided during initiation (anti-CSRF).
      responses:
        '302':
          description: Redirect to frontend `redirect_uri` with success status
        '200':
          description: Tokens stored (pure API context flow, without redirection)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: linked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Identity provider temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /contacts:
    get:
      tags: [contacts]
      summary: List user's Google contacts (proxied)
      description: |
        Proxies `people.connections.list` for the linked user.
        Implements pagination, sorting, and **incremental sync** via `sync_token`.
        **Note**: `person_fields` is *required* by Google, otherwise 400 error from provider.
      operationId: listContacts
      security:
        - bearerAuth: []
      parameters:
        - name: person_fields
          in: query
          required: true
          description: |
            Google People field mask to return (CSV format).
            Ex: `names,emailAddresses,phoneNumbers,photos`.
          schema:
            type: string
            example: names,emailAddresses,phoneNumbers,photos
        - name: page_size
          in: query
          description: Number of items per page (1-200, default 100).
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
        - name: page_token
          in: query
          description: Pagination token (received from `next_page_token`).
          schema:
            type: string
        - name: sort_order
          in: query
          description: Sort order when `sync_token` is not used.
          schema:
            type: string
            enum:
              - LAST_MODIFIED_ASCENDING
              - LAST_MODIFIED_DESCENDING
              - FIRST_NAME_ASCENDING
              - LAST_NAME_ASCENDING
            default: LAST_MODIFIED_ASCENDING
        - name: request_sync_token
          in: query
          description: |
            Return `next_sync_token` on the **last** page.
          schema:
            type: boolean
            default: false
        - name: sync_token
          in: query
          description: |
            Sync token to retrieve only changes since last call.
            **Expires in ~7 days**; if expired, perform a **full sync** without `sync_token`.
          schema:
            type: string
        - name: sources
          in: query
          description: |
            Source mask (Google People `ReadSourceType`). Default: CONTACT + PROFILE.
          schema:
            type: array
            items:
              type: string
              enum:
                - READ_SOURCE_TYPE_CONTACT
                - READ_SOURCE_TYPE_PROFILE
                - READ_SOURCE_TYPE_DOMAIN_PROFILE
                - READ_SOURCE_TYPE_DOMAIN_CONTACT
          example:
            - READ_SOURCE_TYPE_CONTACT
            - READ_SOURCE_TYPE_PROFILE
        - name: q
          in: query
          description: |
            Gateway-side filter (post-processing) on name/email (optional, best effort).
          schema:
            type: string
      responses:
        '200':
          description: Contacts retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsList'
        '400':
          description: Bad request - validation/provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (JWT missing/expired) or Google account not linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient Google scopes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '503':
          description: Provider unavailable / quota exceeded / transient error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with RS256 signature.
        Claims: sub (aircraft_id), aud=skylink, iat, exp (max 15 min)
  schemas:
...
    ContactsList:
      type: object
      description: Paginated result from `people.connections.list` (normalized by gateway).
      required: [items]
      properties:
        items:
          type: array
          description: List of persons, structure matches the `person_fields` mask.
          items:
            $ref: '#/components/schemas/GooglePerson'
        next_page_token:
          type: string
          nullable: true
          description: To be returned in `page_token` for the next page (absent if last page).
          example: "CiQA3P7j…"
        next_sync_token:
          type: string
          nullable: true
          description: To be returned in `sync_token` for incremental sync (present only on the last page if `request_sync_token=true`).
          example: "CPD2tL6i…"
        total_items:
          type: integer
          description: Total number of items (if provided by the provider).
          example: 350
    GooglePerson:
      type: object
      description: Subset of Google `Person` schema, controlled by `person_fields`.
      properties:
        resourceName:
          type: string
          description: Google `people/{id}` identifier.
          example: "people/c1234567890"
        etag:
          type: string
          description: Resource ETag.
        names:
          type: array
          items:
            type: object
            properties:
              displayName:
                type: string
              givenName:
                type: string
              familyName:
                type: string
              unstructuredName:
                type: string
        emailAddresses:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
                format: email
              type:
                type: string
                description: Label (work, home, …)
              metadata:
                type: object
                properties:
                  primary:
                    type: boolean
        phoneNumbers:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              type:
                type: string
              metadata:
                type: object
                properties:
                  primary:
                    type: boolean
        photos:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              default:
                type: boolean
        organizations:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              title:
                type: string
        birthdays:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
                description: Date as provided by Google
        addresses:
          type: array
          items:
            type: object
            properties:
              formattedValue:
                type: string
        metadata:
          type: object
          properties:
            deleted:
              type: boolean
              description: True if the contact has been deleted (incremental sync case).
    Error:
      type: object
      required: [error]
      additionalProperties: false
      description: Standard error envelope (no trace_id).
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                Machine-readable error code.
                Possible values: VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN,
                RATE_LIMIT_EXCEEDED, PROVIDER_UNAVAILABLE, INTERNAL_ERROR
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message.
              example: Invalid input data
            details:
              type: object
              nullable: true
              description: Optional structured details for validation errors.
              properties:
                fields:
                  type: array
                  description: Per-field validation issues
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        description: Affected field name
                        example: lat
                      issue:
                        type: string
                        description: Issue keyword
                        example: range
                      message:
                        type: string
                        description: Human message
                        example: lat must be between -90 and 90
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...