# skylink/models/gateway/contacts_list.py
# coding: utf-8

"""
SkyLink API

SkyLink Connected Aircraft Platform - MVP  Security by Design implementation with JWT authentication, rate limiting, and strict input validation.

The version of the OpenAPI document: 0.6.2
Generated by OpenAPI Generator (https://openapi-generator.tech)

Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations

import json
import pprint
import re  # noqa: F401
from typing import Any, ClassVar, Dict, List, Optional

from pydantic import BaseModel, Field, StrictInt, StrictStr

from skylink.models.gateway.google_person import GooglePerson

try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class ContactsList(BaseModel):
    """
    Paginated result from `people.connections.list` (normalized by the gateway).
    """  # noqa: E501

    items: List[GooglePerson] = Field(
        description="List of persons, structure matches the `person_fields` mask."
    )
    next_page_token: Optional[StrictStr] = Field(
        default=None,
        description="To be returned in `page_token` for the next page (absent if last page).",
    )
    next_sync_token: Optional[StrictStr] = Field(
        default=None,
        description=(
            "To be returned in `sync_token` for incremental sync "
            "(present only on the last page if `request_sync_token=true`)."
        ),
    )
    total_items: Optional[StrictInt] = Field(
        default=None, description="Total number of items (if provided by the provider)."
    )
    __properties: ClassVar[List[str]] = [
        "items",
        "next_page_token",
        "next_sync_token",
        "total_items",
    ]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ContactsList from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={},
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in items (list)
        _items = []
        if self.items:
            for _item in self.items:
                if _item:
                    _items.append(_item.to_dict())
            _dict["items"] = _items
        # set to None if next_page_token (nullable) is None
        # and model_fields_set contains the field
        if self.next_page_token is None and "next_page_token" in self.model_fields_set:
            _dict["next_page_token"] = None

        # set to None if next_sync_token (nullable) is None
        # and model_fields_set contains the field
        if self.next_sync_token is None and "next_sync_token" in self.model_fields_set:
            _dict["next_sync_token"] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ContactsList from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate(
            {
                "items": (
                    [GooglePerson.from_dict(_item) for _item in obj.get("items")]
                    if obj.get("items") is not None
                    else None
                ),
                "next_page_token": obj.get("next_page_token"),
                "next_sync_token": obj.get("next_sync_token"),
                "total_items": obj.get("total_items"),
            }
        )
        return _obj